Fix all remaining compiler warnings, enable -Werror in CI
Full clean build previously had 205 warnings, all -Wunused-parameter,
all in virtual listener interfaces (RestApiListener, RestEndpointListener,
WebsocketApiListener, WebsocketMessageHandler) with empty default {}
bodies - parameter names document the interface but the no-op default
never uses them. Fixed by dropping the names on the ~90 affected
declarations (standard C++ idiom - unnamed parameters never trigger
this warning).

Also found and fixed while cleaning this up:
- Signing::signUserSignedAction took an isMainnet parameter its own
  body never used (the mainnet/testnet distinction is already baked
  into the signed action's hyperliquidChain field by the caller, and
  the EIP-712 domain's chain id is a fixed constant regardless of
  environment - confirmed by grepping every call site). Since this is
  an internal (src/) header, not part of the public API, removed the
  parameter entirely rather than just silencing it - updated all 10
  call sites (2 in Signing.cpp itself, 8 across 3 test files).
- rest_twap_slices.cpp compared a uint64_t twapId against
  static_cast<int>(*twapResp.twapId) - a lossy narrowing cast that
  still triggered -Wsign-compare on the final comparison. Removed the
  pointless cast; both sides are uint64_t.
- 9 examples had an onDisconnected(bool hasError, const std::string&
  errMsg) override whose body ignored both parameters - now logs them
  (more informative for anyone learning from these examples than
  just silencing the warning).

Extended -Wall -Wextra -Wpedantic (previously library-only) to tests/
and examples/ targets too, via a shared HYPERLIQUID_WARNING_FLAGS
CMake variable - this is what surfaced the two real bugs above,
previously invisible since those targets weren't checked at all.

Added an opt-in HYPERLIQUID_WARNINGS_AS_ERRORS CMake option (off by
default, so local dev builds stay permissive mid-edit) and turned it
on in CI specifically, so new warnings can't silently creep back in.
Verified the option actually works both ways: confirmed a clean build
passes with it on, and confirmed a deliberately-injected warning fails
the build with it on (reverted before committing).

Includes the same one-line dangling-reference fix as #67 (unmerged
at time of writing) - this branch needed it too to reach zero
warnings; will no-op cleanly on rebase once #67 merges.

+++ examples/rest_twap_slices.cpp
@@ -49,7 +49,7 @@ int main()
         spdlog::info("[poll {}] userTwapSliceFills: {} total fills", i + 1, slices.fills.size());
         for (const auto& slice : slices.fills)
         {
-            if (slice.twapId != static_cast<int>(*twapResp.twapId)) continue;
+            if (slice.twapId != *twapResp.twapId) continue;
             spdlog::info("  MATCH twapId={} coin={} px={} sz={} time={}",
                          slice.twapId, slice.fill.coin, slice.fill.px, slice.fill.sz, slice.fill.time);
         }
